Add usage Channels, change token lifetime
Add usage Django Channels. If WebSocket connection is established,
the user will receive a notification of enrollment in the course as
a student or teacher.
Authentication possible with AuthMiddlewareStack or by passing the
"token" parameter when creating a WebSocket connection.

Add pagination when requesting big data.

Change access token lifetime to 40 minutes.